requirements.txt
streamlit
google-generativeai
python-dotenv
pyrebase4
transformers
matplotlib
reportlab
